+++ lib/presentation/screen/profile/about_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:auto_route/auto_route.dart';
+
+@RoutePage()
+class AboutPage extends StatelessWidget {
+  const AboutPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('关于我们')),
+      body: const Center(
+        child: Text('这里是关于我们页面', style: TextStyle(fontSize: 24)),
+      ),
+    );
+  }
+}
