+++ lib/presentation/screen/profile/settings_screen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:auto_route/auto_route.dart';
+
+@RoutePage()
+class SettingsPage extends StatelessWidget {
+  const SettingsPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('设置')),
+      body: const Center(
+        child: Text('这里是设置页面', style: TextStyle(fontSize: 24)),
+      ),
+    );
+  }
+}
